feat: add DefaultStickyRightSide for Layout + reorder files
