+++ client/src/components/PostsView.tsx
@@ -1,5 +1,6 @@
 import { FC, HTMLAttributes } from 'react';
 import Post from './Post';
+import { mergeClassNames } from '../utils';
 
 interface Props extends HTMLAttributes<HTMLDivElement> {
   data: any; // TODO: add real type
@@ -9,7 +10,7 @@ const PostsView: FC<Props> = ({ ...props }) => {
   const posts = [1, 2, 3, 4, 5, 6];
 
   return (
-    <div {...props} className={'w-full' + ` ${props.className}`}>
+    <div {...props} className={mergeClassNames('w-full', props.className)}>
       {posts.map((v) => {
         return <Post key={v} />;
       })}
