+++ client/src/hooks/useMaxScroll.ts
@@ -1,1 +1,1 @@
-export { };
+export {};
